Addressing some memory leaks:
Moved ConnectionLiveData to activity with data connection status being saved to SharedPreferences for use in other fragments.

Moved auto navigation from LoginFragment to ToDoFragment from MainActivity to LoginFragment as we can easily ensure view is not built unless required.

AlertDialog in TodoFragment is no longer injected as it's only used once and is not being disposed of properly.